Refactor project modules
